add a streaming API
The Go standard library only supports en/decryption
of chunked messages. In general it is recommended to design your
protocol/appliction to use message based encryption.

Unfortunately it is not always possible to do so. In those situations
it is nice to have a library which provides a streaming API.
Instead of encrypting your message directly the streaming API wraps an
io.Reader or an io.Writer and en/decrypts the input of the reader/writer.